refactor: iOS 소모품 등록 화면 정리
