+++ iosApp/iosApp/App/ContentView.swift
@@ -3,6 +3,7 @@ import SwiftUI
 struct ContentView: View {
     var body: some View {
         OBRitNavigation()
+            .dismissKeyboardOnBackgroundTap()
     }
 }
